Refactor backend: modularize models and routes, add .env.example, update README
