+++ .env.example
@@ -1,13 +1,8 @@
-# OpenAI Configuration
-OPENAI_API_KEY=your_openai_api_key_here
-
-# MongoDB Configuration
+# MongoDB connection URI
 MONGODB_URI=mongodb://localhost:27017/research_reader
 
-# Server Configuration
-PORT=5000
-NODE_ENV=development
+# OpenAI API Key
+OPENAI_API_KEY=your_openai_api_key_here
 
-# File Upload Configuration
-MAX_FILE_SIZE=10485760  # 10MB in bytes
-UPLOAD_DIR=uploads/papers
+# Server port
+PORT=5000
