+++ README.md
@@ -1,118 +1,46 @@
- 
+# Research Reader
 
-```markdown
-# Cross-Platform Mobile LLM-Powered Research Reader
+## Project Structure
 
-**Objective**  
-Develop a mobile app enabling users to read, summarize, and interact with research papers via LLM APIs, with cross-platform support through Flutter.
+- `Node.js` - Main backend server (now uses modular routes and models)
+- `models/` - Mongoose models for Paper, Note, Chat
+- `routes/` - Express routers for papers, notes, chat
+- `main.tsx` - Main frontend React app
+- `.env.example` - Example environment variables
 
-## Project Overview
+## Setup
 
-### Core Features
-- **PDF/Text Upload & Viewing**  
-  Support for document uploads with preview functionality
-- **LLM Integration**  
-  Summarization, Q&A, and content analysis via OpenAI/Anthropic APIs
-- **Annotation Tools**  
-  Note-taking, highlighting, and tagging for active reading
-- **User Management**  
-  Authentication, reading history, and personalized settings
+1. Copy `.env.example` to `.env` and fill in your values.
+2. Install dependencies:
+   ```bash
+   npm install
+   ```
+3. Start the backend:
+   ```bash
+   node Node.js
+   ```
+4. Start the frontend (if separate):
+   ```bash
+   # e.g., npm start or yarn start
+   ```
 
-### Architecture
+## Environment Variables
 
-```
-graph TD
-    A[Mobile App (Flutter)] --> B[UI Layer]
-    A --> C[Business Logic]
-    A --> D[Network Layer]
-    D --> E[Backend API (Node.js/FastAPI)]
-    E --> F[Auth/File Upload]
-    E --> G[LLM API Integration]
-    G --> H[Cloud-Based LLM]
-    E --> I[Database/Storage (Firebase/S3/MongoDB)]
-```
+See `.env.example` for required variables:
+- `MONGODB_URI` - MongoDB connection string
+- `OPENAI_API_KEY` - Your OpenAI API key
+- `PORT` - Server port (default 5000)
 
-## Development Roadmap
+## Backend Refactor
 
-| Phase              | Timeline       | Key Activities                          |
-|---------------------|----------------|-----------------------------------------|
-| Planning & Setup    | Week 1         | Feature definition, wireframes, repo setup |
-| Core Development    | Weeks 2–5      | UI/Backend/API/LLM integration           |
-| Advanced Features   | Weeks 6–7      | Auth, notes, UI optimization             |
-| Testing/Deployment  | Weeks 8–10     | Multi-device testing, app store submission |
+- Models and routes are now modular for maintainability.
+- All API endpoints are mounted under `/api`.
+- Sensitive config is loaded from environment variables.
 
-**Alternative Frameworks**  
-Consider these if technical needs change:
-- **React Native** (JavaScript/TypeScript)
-- **Xamarin/.NET MAUI** (C#)
-- **Kotlin Multiplatform** (Kotlin)
+## Contributing
 
-## Setup & Installation
+Pull requests are welcome. For major changes, please open an issue first to discuss what you would like to change.
 
-**Prerequisites**  
-```
-# Clone repository
-git clone [REPO_URL]
+## License
 
-# Initialize submodules
-git submodule init && git submodule update
-```
-
-**Environment Setup**  
-| Service       | Command                          |
-|---------------|----------------------------------|
-| Flutter       | `flutter pub get`               |
-| Node.js       | `npm install`                   |
-| FastAPI       | `pip install fastapi uvicorn`   |
-
-**Environment Variables**  
-Create `.env` file with:
-```
-OPENAI_API_KEY=your_key_here
-FIREBASE_PROJECT_ID=your_project_id
-STORAGE_BUCKET_NAME=your_bucket_name
-```
-
-## Next Steps (Week 1)
-
-**1. Feature Prioritization**  
-- Create user stories for core functionality:  
-  - PDF upload validation & preview  
-  - Basic LLM integration (text extraction → summarization)  
-  - Local storage for temporary files
-
-**2. Wireframe Development**  
-- Tools: Figma, Adobe XD, or Excalidraw  
-- Focus on key workflows:  
-  - Document upload → processing → summary display  
-  - Annotation → note management → export
-
-**3. Project Structure Setup**  
-```
-# Recommended folder structure
-frontend/
-├── lib/
-│   └── widgets/
-├── models/
-└── services/
-backend/
-├── routes/
-├── controllers/
-└── utils/
-docs/
-├── architecture/
-└── api-docs/
-```
-
-**4. CI/CD Pipeline**  
-- Set up GitHub Actions for:  
-  - Flutter build validation  
-  - Backend API tests  
-  - Environment variable encryption
-
-**5. Documentation Templates**  
-- Create boilerplate for:  
-  - API documentation (Swagger/OpenAPI)  
-  - Feature requirements (Aha!-style templates)  
-  - Testing scenarios (Postman collections)
-```
+[MIT](LICENSE)
