+++ models/Note.js
@@ -1,0 +1,11 @@
+const mongoose = require('mongoose');
+
+const noteSchema = new mongoose.Schema({
+  paperId: { type: mongoose.Schema.Types.ObjectId, ref: 'Paper' },
+  content: String,
+  userId: String,
+  createdAt: { type: Date, default: Date.now },
+  updatedAt: { type: Date, default: Date.now },
+});
+
+module.exports = mongoose.model('Note', noteSchema);
